move token escaper
